// src/network/messages/message-type.ts
enum MessageType {
  APPEND_CHILD_TO_CONTAINER = 'APPEND_CHILD_TO_CONTAINER',
  APPEND_CHILD = 'APPEND_CHILD',
  APPEND_INITIAL_CHILD = 'APPEND_INITIAL_CHILD',
  CLEAR_CONTAINER = 'CLEAR_CONTAINER',
  COMMIT_UPDATE = 'COMMIT_UPDATE',
  CREATE_INSTANCE = 'CREATE_INSTANCE',
  HIDE_INSTANCE = 'HIDE_INSTANCE',
  INSERT_BEFORE = 'INSERT_BEFORE',
  INSERT_IN_CONTAINER_BEFORE = 'INSERT_IN_CONTAINER_BEFORE',
  INVOKE_CALLBACK = 'INVOKE_CALLBACK',
  REMOVE_CHILD_FROM_CONTAINER = 'REMOVE_CHILD_FROM_CONTAINER',
  REMOVE_CHILD = 'REMOVE_CHILD',
  UNHIDE_INSTANCE = 'UNHIDE_INSTANCE',
}

export default MessageType